+++ extensions/ql-vscode/src/common/octokit.ts
@@ -1,10 +1,13 @@
 import { Octokit } from "@octokit/rest";
 import { retry } from "@octokit/plugin-retry";
-import fetch from "node-fetch";
 
 export const AppOctokit = Octokit.defaults({
   request: {
-    fetch,
+    // MSW replaces the global fetch object, so we can't just pass a reference to the
+    // fetch object at initialization time. Instead, we pass a function that will
+    // always call the global fetch object.
+    fetch: (input: string | URL | Request, init?: RequestInit) =>
+      fetch(input, init),
   },
   retry,
 });