+++ extensions/ql-vscode/test/unit-tests/variant-analysis/gh-api/gh-api-client.test.ts
@@ -13,7 +13,7 @@ import { response as variantAnalysisRepoJson_response } from "../../../../src/co
 import { testCredentialsWithRealOctokit } from "../../../factories/authentication";
 
 const mockServer = new MockGitHubApiServer();
-beforeAll(() => mockServer.startServer());
+beforeAll(() => mockServer.startServer("error"));
 afterEach(() => mockServer.unloadScenario());
 afterAll(() => mockServer.stopServer());